requirements.txt
Django==3.2.15
django-extensions==3.2.0
djangorestframework==3.13.1
